Clamp the scaling factor to reasonable values
Under specific scheduling conditions, e.g. while running in an overloaded VM, we
might end up with really bad scaling values:

   Over 1014886 us, ratio=100.44%, scaling   inf%, 36 loops, 500 MB memory
   Over 1024451 us, ratio=100.21%, scaling  -inf%, 29 loops, 500 MB memory
   Over 1014840 us, ratio=100.55%, scaling   inf%, 25 loops, 500 MB memory

Add some code to avoid that condition, and also to move the needle a little more
slowly, so that a transient spike will not result in a completely bogus scaling
factor value.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>